Update TUNNA
Add some clarity to the status of a nonprofit to delete references to retains and patronage, and clarify LOOT quit as claiming back "restricted" donations.  This would be similar to making a restricted donation that is held in the smart contract so long as the person making the contribution agrees with the direction of the nonprofit.